Prereqs for callback transfer
VF callback plugins are being moved from validations-common repo
to the validations-libs repo.

During the transitionary period, before the changes to both upstream repos
are fully merged, the callback might be present in the source
used for package build. In the case they are, they should be used.
Otherwise, an empty directory will be created, effectivelly mocking
the presence of callbacks.

Depends-On: https://review.opendev.org/c/openstack/validations-libs/+/820551

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I0629898c95416b211384119580f5b0a8ba6d1765