Man pages for validation framework
Documentation in a form of UNIX man pages is compiled and installed with the package
and organized in two parts:

1. vf.1 - describing the CLI of the Validation Framework runtime
2. validations-libs.3 - describing the API of the Validation Framework

Both are installed in the relevant man directories and entered into the database.

Additional sphinx-build dependencies were introduced to enable documentation
build process and allow for automated API documentation.

Finally, installation of the legacy documentation, in a form of README document,
was clarified.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I876ee81e9a70049d38d0f36916e56b39a851afa4